Reorganize the locations tests and factories. Remove roughly duplicated factory.
